requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-writable-nested==0.5.4
phonenumbers==8.11.2
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
gunicorn==20.0.4

